Revert to 'ptrdiff_t' anyway.
Too bad that 'splint' now warns about '%td'.

Let's see what the different C compilers will tell. ;o)

+++ ctangle.c
@@ -1634,14 +1634,14 @@ default:err_print("! Double @ should be used in limbo");
 void
 print_stats(void){
 puts("\nMemory usage statistics:");
-printf("%ld names (out of %ld)\n",
-(long)(name_ptr-name_dir),(long)max_names);
-printf("%ld replacement texts (out of %ld)\n",
-(long)(text_ptr-text_info),(long)max_texts);
-printf("%ld bytes (out of %ld)\n",
-(long)(byte_ptr-byte_mem),(long)max_bytes);
-printf("%ld tokens (out of %ld)\n",
-(long)(tok_ptr-tok_mem),(long)max_toks);
+printf("%td names (out of %ld)\n",
+(ptrdiff_t)(name_ptr-name_dir),(long)max_names);
+printf("%td replacement texts (out of %ld)\n",
+(ptrdiff_t)(text_ptr-text_info),(long)max_texts);
+printf("%td bytes (out of %ld)\n",
+(ptrdiff_t)(byte_ptr-byte_mem),(long)max_bytes);
+printf("%td tokens (out of %ld)\n",
+(ptrdiff_t)(tok_ptr-tok_mem),(long)max_toks);
 }
 
 /*:103*/
